feat: 상품 패널에 대한 컴포넌트 구현
