+++ src/refactoring/entities/coupon/ui/CouponProvider.tsx
@@ -1,21 +1,7 @@
-import { createContext, ReactNode, useContext, useMemo } from 'react';
+import { ReactNode, useMemo } from 'react';
 import { Coupon } from '../../../../types';
 import { useCoupons } from '../model/useCoupon';
-
-interface CouponContextType {
-  coupons: Coupon[];
-  addCoupon: (coupon: Coupon) => void;
-}
-
-const CouponContext = createContext<CouponContextType | null>(null);
-
-export function useCouponContext() {
-  const context = useContext(CouponContext);
-  if (!context) {
-    throw new Error('useCouponStore must be used within a CouponProvider');
-  }
-  return context;
-}
+import { CouponContext } from '../model/useCouponContext';
 
 export function CouponProvider({
   children,
